+++ generators/typescript/utils/contexts/src/sdk-context/sdk-client-class/GeneratedEndpointImplementation.ts
@@ -1,0 +1,23 @@
+import { ExampleEndpointCall, HttpEndpoint } from "@fern-fern/ir-sdk/api";
+import { GetReferenceOpts } from "@fern-typescript/commons";
+import { OptionalKind, ParameterDeclarationStructure, ts } from "ts-morph";
+import { SdkContext } from "../SdkContext";
+
+export interface EndpointSignature {
+    parameters: OptionalKind<ParameterDeclarationStructure & { docs?: string }>[];
+    returnTypeWithoutPromise: ts.TypeNode;
+}
+
+export interface GeneratedEndpointImplementation {
+    endpoint: HttpEndpoint;
+    getStatements: (context: SdkContext) => ts.Statement[];
+    getOverloads: (context: SdkContext) => EndpointSignature[];
+    getSignature: (context: SdkContext) => EndpointSignature;
+    getDocs: (context: SdkContext) => string | undefined;
+    getExample: (args: {
+        context: SdkContext;
+        example: ExampleEndpointCall;
+        opts: GetReferenceOpts;
+        clientReference: ts.Identifier;
+    }) => ts.Expression | undefined;
+}