+++ packages/cli/openapi-ir-to-fern/src/__test__/x-fern-global-headers.test.ts
@@ -1,0 +1,5 @@
+import { testConvertOpenAPI } from "./testConvertOpenApi";
+
+describe("x-fern-global-headers", () => {
+    testConvertOpenAPI("x-fern-global-headers", "openapi.yml");
+});