+++ packages/seed/src/cli.ts
@@ -61,6 +61,11 @@ function addTestCommand(cli: Argv) {
                     description: "Determines whether or not snapshots are written to disk",
                     default: false
                 })
+                .option("skip-scripts", {
+                    type: "boolean",
+                    demandOption: false,
+                    default: false
+                })
                 .option("log-level", {
                     default: LogLevel.Info,
                     choices: LOG_LEVELS
@@ -100,7 +105,8 @@ function addTestCommand(cli: Argv) {
                         docker: parsedDockerImage,
                         logLevel: argv["log-level"],
                         numDockers: argv.parallel,
-                        keepDocker: argv.keepDocker
+                        keepDocker: argv.keepDocker,
+                        skipScripts: argv.skipScripts
                     });
                 } else {
                     await testWorkspaceFixtures({
@@ -113,7 +119,8 @@ function addTestCommand(cli: Argv) {
                         logLevel: argv["log-level"],
                         numDockers: argv.parallel,
                         taskContextFactory,
-                        keepDocker: argv.keepDocker
+                        keepDocker: argv.keepDocker,
+                        skipScripts: argv.skipScripts
                     });
                 }
             }