+++ seed/java-sdk/exhaustive/local-files/resources/endpoints/httpmethods/HttpMethodsClient.java
@@ -18,6 +18,7 @@ import java.lang.RuntimeException;
 import java.lang.String;
 import okhttp3.Headers;
 import okhttp3.HttpUrl;
+import okhttp3.OkHttpClient;
 import okhttp3.Request;
 import okhttp3.RequestBody;
 import okhttp3.Response;
@@ -46,7 +47,11 @@ public class HttpMethodsClient {
       .addHeader("Content-Type", "application/json")
       .build();
     try {
-      Response response = clientOptions.httpClient().newCall(okhttpRequest).execute();
+      OkHttpClient client = clientOptions.httpClient();
+      if (requestOptions.getTimeout().isPresent()) {
+        client = client.newBuilder().readTimeout(requestOptions.getTimeout().get(), requestOptions.getTimeoutTimeUnit()).build();
+      }
+      Response response = client.newCall(okhttpRequest).execute();
       if (response.isSuccessful()) {
         return ObjectMappers.JSON_MAPPER.readValue(response.body().string(), String.class);
       }
@@ -81,7 +86,11 @@ public class HttpMethodsClient {
       .addHeader("Content-Type", "application/json")
       .build();
     try {
-      Response response = clientOptions.httpClient().newCall(okhttpRequest).execute();
+      OkHttpClient client = clientOptions.httpClient();
+      if (requestOptions.getTimeout().isPresent()) {
+        client = client.newBuilder().readTimeout(requestOptions.getTimeout().get(), requestOptions.getTimeoutTimeUnit()).build();
+      }
+      Response response = client.newCall(okhttpRequest).execute();
       if (response.isSuccessful()) {
         return ObjectMappers.JSON_MAPPER.readValue(response.body().string(), ObjectWithOptionalField.class);
       }
@@ -117,7 +126,11 @@ public class HttpMethodsClient {
       .addHeader("Content-Type", "application/json")
       .build();
     try {
-      Response response = clientOptions.httpClient().newCall(okhttpRequest).execute();
+      OkHttpClient client = clientOptions.httpClient();
+      if (requestOptions.getTimeout().isPresent()) {
+        client = client.newBuilder().readTimeout(requestOptions.getTimeout().get(), requestOptions.getTimeoutTimeUnit()).build();
+      }
+      Response response = client.newCall(okhttpRequest).execute();
       if (response.isSuccessful()) {
         return ObjectMappers.JSON_MAPPER.readValue(response.body().string(), ObjectWithOptionalField.class);
       }
@@ -157,7 +170,11 @@ public class HttpMethodsClient {
       .addHeader("Content-Type", "application/json")
       .build();
     try {
-      Response response = clientOptions.httpClient().newCall(okhttpRequest).execute();
+      OkHttpClient client = clientOptions.httpClient();
+      if (requestOptions.getTimeout().isPresent()) {
+        client = client.newBuilder().readTimeout(requestOptions.getTimeout().get(), requestOptions.getTimeoutTimeUnit()).build();
+      }
+      Response response = client.newCall(okhttpRequest).execute();
       if (response.isSuccessful()) {
         return ObjectMappers.JSON_MAPPER.readValue(response.body().string(), ObjectWithOptionalField.class);
       }
@@ -185,7 +202,11 @@ public class HttpMethodsClient {
       .addHeader("Content-Type", "application/json")
       .build();
     try {
-      Response response = clientOptions.httpClient().newCall(okhttpRequest).execute();
+      OkHttpClient client = clientOptions.httpClient();
+      if (requestOptions.getTimeout().isPresent()) {
+        client = client.newBuilder().readTimeout(requestOptions.getTimeout().get(), requestOptions.getTimeoutTimeUnit()).build();
+      }
+      Response response = client.newCall(okhttpRequest).execute();
       if (response.isSuccessful()) {
         return ObjectMappers.JSON_MAPPER.readValue(response.body().string(), boolean.class);
       }