+++ seed/ts-sdk/examples/examples-with-api-reference/src/api/resources/index.ts
@@ -1,0 +1,8 @@
+export * as commons from "./commons";
+export * as file from "./file";
+export * as types from "./types";
+export * from "./types/types";
+export * from "./types/errors";
+export * as health from "./health";
+export * as service from "./service";
+export * from "./service/client/requests";